+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/KotlinGradleExtensionTest.java
@@ -74,6 +74,24 @@ public class KotlinGradleExtensionTest extends GradleIntegrationTest {
 	}
 
 	@Test
+	public void integration_shyiko() throws IOException {
+		setFile("build.gradle").toLines(
+				"plugins {",
+				"    id 'nebula.kotlin' version '1.0.6'",
+				"    id 'com.diffplug.gradle.spotless'",
+				"}",
+				"repositories { mavenCentral() }",
+				"spotless {",
+				"    kotlinGradle {",
+				"        ktlint('0.21.0')",
+				"    }",
+				"}");
+		setFile("configuration.gradle.kts").toResource("kotlin/ktlint/basic.dirty");
+		gradleRunner().withArguments("spotlessApply").build();
+		assertFile("configuration.gradle.kts").sameAsResource("kotlin/ktlint/basic.clean");
+	}
+
+	@Test
 	public void indentStep() throws IOException {
 		setFile("build.gradle").toLines(
 				"plugins {",